requirements.txt
# ask-openai
openai
keyring
keyrings.cryptfile

# wcl function (wes clone)
giturlparse
